requirements.txt
Flask==2.0.2
redis==4.1.3
websocket-client==1.2.3
requests==2.27.1
celery==5.2.3